Modified conv_util.computeConv2DInfo to calculate the convolution 2d info when using atrous convolution and dilation rates
See https://www.tensorflow.org/api_docs/python/tf/nn/atrous_conv2d

> Atrous convolution is equivalent to standard convolution with upsampled
> filters with effective height `filter_height + (filter_height - 1) * (rate - 1)`
> and effective width `filter_width + (filter_width - 1) * (rate - 1)`,
> produced by inserting `rate - 1` zeros along consecutive elements across
> the filters' spatial dimensions.

When there is a dilation rate, this converts a filter dimension to the
effective filter dimension, so it can be used in a standard convolution.
This modification uses the above logic.

Changed rates to dilations everywhere, to make consistent with
Tensorflow api