Implemented atrous 1d and 2d convolution.
* Updated backend gpu and cpu to do atrous convolution when there is a dilation rate.
* Refactored backend_cpu.conv2d to be consistent with backend_gpu.conv2d, and also to make the way it does dilated convolution consistent.
* Added tests for 1d and 2d convolution with dilation rates that show the effect on the filter when dilation rates are set.
* Updated computeDefaultPad to account for a dilation rate.

Implemented atrous convolution for depthwiseConv2D

* Modified cpu depthwiseConv2D logic to be similar to that on the GPU,
so that dilation can be easily applied.

Still need to add more tests for depthwise atrous convolution

Per feedback, changed order of parameters to match
Tensorflow api.  Added dataFormat parameter to conv1d, conv2d, and
depthwiseConv2d, but did not yet implement that parameter; it defaults
to a value and cannot be a different value until the functionality
is implemented

raising error when gradient is done with atrous convolution