[io] Fix a bug in concatenateTypedArrays
Previously, it was incorrectly assumed that the `buffer` of a
`TypedArray` has the same `byteLength` as the `TypedArray` itself.
The two can actually be different if the `TypedArray` is created
from an offset and a length from an `ArrayBuffer`, e.g.,
`const x = new Float32Array(myBuffer, 32, 4)`.
This actually happens during the desrialization of a `tf.Model`.

This PR fixes it by checking whether the `byteLength` and
`buffer.byteLength` of a `TypedArray` are equal, and if not, creating
a new `TypedArray` with the buffer that matches the actual `byteLength`.

BUG