"Low-Level" QR Decomposition, bandPart, triangularSolve
The QR Decomposition implementation uses Givens Rotations for
elimination. There's a few reasons, why the Givens Rotation was
chosen over the Householder method:
  * During a quick trial, there seemed to be no practical
    performance difference in JS between Householder and Givens.
    JS itself seems to introduce enough overhead to make the
    fewe extra FLOPS irrelevant.
  * Givens Method is easier to implement numerically stable
    (e.g. there is no underflow-safe norm necessary).
  * Givens Method is easier to backpropagate.
  * Givens Method ensure det(Q) = 1
  * Givens Rotations seem to be smoother when it comes to Pertubation,
    resulting in smoother gradients.
  * Givens Method is easier to parallelize

As long as R is non-singular there is always a way to produce a
canonical representation from a QR Decompostion (e.g. make the
diagonal of R all positive). That also means that there is no
compatibilty issue with whichever QR implementation Tensorflow
for Python/C/C++ chooses.

Both `bandPart` and `triangularSolve` were necessary to implement
the symbolic backpropagation of the QR Decomposition.